Create initial functionality
This code was written while on a plane, without WiFi.  I was able to get
as far as I did because of yarn’s `--offline` mode, but there is a lot
more to do, like adding CI, creating a Github repo, adding linting,
and even checking if the name is available on the npm registry.

All that said, this _is_ working so far against my branch of ESLint 
which implements the ability to provide a function as the `fix` parameter
when configuring a CLIEngine.