Simplify API
This creates a new exported function, `fix`, which performs most of
the work that was being done previously in the `cli.js` consumer.

This will hopefully make it easier to use eslint-filtered-fix, as now
it’s only necessary to provide a single function a file or list of
files, a set of fix options, and any eslint options if necessary.