Breaking: Support ESLint 8 (#19)
This is a breaking change to support the new `ESLint` class that replaces `CLIEngine`.  Version 8.0.0 of eslint no longer includes `CLIEngine` (https://eslint.org/docs/user-guide/migrating-to-8.0.0#remove-cliengine), so these changes are necessary to support version 8.x.  The new class uses async methods, which means that the exported `fix` function from this package is now async as well, which is the major breaking change.

Also, the result of running `fix()` is now an array of results, and no longer contains roll-up statistics such as the total number of errors and warnings.