Improve performance by adding a rule filter to ensure that only focused rules are evaluated in lint (#27)
I was trying to improve the speed of eslint-nibble in certain workflows -- in this case, notably, while autofixing fixable lint infractions. I wrapped each statement of the fix function here with a console.time/timeEnd including the full function. As we can see, the `calculateFixes` call is the main contributor to the total time:

```
{ options: { cache: true, fix: [Function (anonymous)] } }
makeFixer: 0.121ms
getEslintCli: 46.971ms
calculateFixes: 49.521s
applyFixes: 0.598ms
calculateFixes 2: 44.243s
fix: 1:33.813 (m:ss.mmm)
```

Assumedly the reason is due to needing to run every rule on every file -- even though we've already cut down the rules that we care about.

Obviously we can't statically override rule configs as they are dynamic and recursively defined; however, we can add a rule filter to turn off any rules we do not care about.

```
{
  options: {
    cache: true,
    fix: [Function (anonymous)],
    ruleFilter: [Function: ruleFilter]
  }
}
makeFixer: 0.2ms
getEslintCli: 47.422ms
calculateFixes: 2.153s
applyFixes: 0.195ms
calculateFixes 2: 140.463ms
fix: 2.342s
```

On a fairly large project with a robust ruleset, this reduced the inclusive time of the fix function from 1m 46s to 2.3s. I can't think of any issues this would create, but please review and let me know if I'm crazy here!

Edit -- found an issue which was that the config setting was added in eslint v9, so... now I'm checking for that.